backend/pyproject.toml
[project]
name = "tata-motors-crm-backend"
version = "1.0.0"
description = "Production CRM for Tata Motors Authorized Dealership – Jaipur"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "redis>=5.0.4",
  "httpx>=0.27.0",
  "pydantic-settings>=2.3.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  "aiofiles>=23.2.1",
  "pillow>=10.4.0",
  "python-dateutil>=2.9.0",
  "jinja2>=3.1.4",
  "weasyprint>=62.3",
  "celery>=5.4.0",
  "flower>=2.0.1",
  "boto3>=1.34.0",
  "sendgrid>=6.11.0",
  "twilio>=9.2.0",
  "qrcode>=7.4.2",
  "reportlab>=4.2.2",
  "openpyxl>=3.1.5"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.6",
  "httpx>=0.27.0",
  "ruff>=0.4.5"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
